consolidate ColorInputSet for shared use between Text and BarCode
