+++ packages/ui/src/components/Designer/Sidebar/DetailView/BarCodePropEditor.tsx
@@ -1,4 +1,4 @@
-import React, { useContext } from 'react';
+import React, { CSSProperties, useContext } from 'react';
 import {
   SchemaForUI,
   DEFAULT_FONT_COLOR,
@@ -7,64 +7,21 @@ import {
 } from '@pdfme/common';
 import { I18nContext } from '../../../../contexts';
 import { SidebarProps } from '..';
-import { XMarkIcon } from '@heroicons/react/24/outline';
-
-const inputStyle = {
-  width: '90%',
-  color: '#333',
-  background: 'none',
-  borderRadius: 2,
-  border: '1px solid #767676',
-};
-
-const ColorInputSet = (props: {
-  label: string;
-  value: string;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  onClear: () => void;
-}) => {
-  const { label, value, onChange, onClear } = props;
-  const formattedLabel = label.replace(/\s/g, '');
-
-  return (
-    <div style={{ width: '45%', marginBottom: '10px' }}>
-      <label htmlFor={`input-${formattedLabel}`}>{label}</label>
-      <div style={{ display: 'flex' }}>
-        <input
-          id={`input-${formattedLabel}`}
-          name={`input-${formattedLabel}`}
-          onChange={onChange}
-          value={value}
-          type="color"
-          style={inputStyle}
-        />
-        <button
-          onClick={onClear}
-          style={{
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            background: 'none',
-            borderRadius: 2,
-            border: '1px solid #767676',
-            cursor: 'pointer',
-          }}
-        >
-          <XMarkIcon width={10} height={10} />
-        </button>
-      </div>
-    </div>
-  );
-};
+import ColorInputSet from './FormComponents/ColorInputSet';
 
 const BarcodePropEditor = (
   props: Pick<SidebarProps, 'changeSchemas'> & { activeSchema: SchemaForUI }
 ) => {
   const i18n = useContext(I18nContext);
   const { changeSchemas, activeSchema } = props;
-  const barcodeHasText = activeSchema.type !== 'qrcode' && activeSchema.type !== 'gs1datamatrix';
 
   if (activeSchema.type === 'text' || activeSchema.type === 'image') return <></>;
+
+  const barcodeHasText = activeSchema.type !== 'qrcode' && activeSchema.type !== 'gs1datamatrix';
+  const colorInputStyles: CSSProperties = {
+    width: barcodeHasText ? '30%' : '45%',
+    marginBottom: '10px',
+  };
 
   return (
     <section style={{ fontSize: '0.7rem' }}>
@@ -88,6 +45,7 @@ const BarcodePropEditor = (
               { key: 'barcolor', value: DEFAULT_FONT_COLOR, schemaId: activeSchema.id },
             ])
           }
+          extraStyle={colorInputStyles}
         />
 
         <ColorInputSet
@@ -99,6 +57,7 @@ const BarcodePropEditor = (
           onClear={() => {
             changeSchemas([{ key: 'backgroundcolor', value: '', schemaId: activeSchema.id }])
           }}
+          extraStyle={colorInputStyles}
         />
 
         {barcodeHasText && (
@@ -111,6 +70,7 @@ const BarcodePropEditor = (
             onClear={() => {
               changeSchemas([{ key: 'textcolor', value: '', schemaId: activeSchema.id }])
             }}
+            extraStyle={colorInputStyles}
           />
         )}
       </div>
